upgrade genetic algo: added gradient descent, and hyperparameter "desperation"
Training AI models gives a lot of time to throw the kitchen sink at the alternate method!  Added a per-generation gradient descent step that gets called when there was no fitness improvement in a generation. When even that starts to fail, a new hyper-parameter called "desperation" gets incremented. It increases the chance of mutants as well as the extent to which they may be mutated.  It helps a little, and it's fun to watch that fitness graph converge that tiny bit more.