still scratching head over genetic algo
For some reason, after the genetic algorithm is ending, having found a really good fit, the result gets somehow reverted to a less-good fit, seemingly between the end of the GA and the result getting saved to disk. Feels like a pass-by-reference thing, so trying deepcopy() in various places. Sometimes, the result looks a lot like the initial estimate, even though that should be long-gone after 500 generations! Other times, the result looks like something near the end, a good fit, but not the best one supposedly just achieved.