GA improved by making it a global
Passing the GA instance though all these functions was hard to debug. Since we only ever need one instance of it at a time, I just make it a global. This seems to have resolved the weirdness where best result was getting 'replaced' by an earlier, non-best result after the GA and before the disk write.